requirements.txt
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
bangla-pdf-ocr>=0.1.0
python-docx==1.1.2
